change int in double for created_at and expires in

+++ swifty-companion/APIModel.swift
@@ -11,8 +11,8 @@ import Foundation
 
 struct TokenAPI: Codable {
     var access_token: String?
-    var created_at: Int
-    var expires_in: Int
+    var created_at: Double
+    var expires_in: Double
 }
